shorted integration time in test
