+++ .github/workflows/docs.yml
@@ -5,7 +5,7 @@ on:
   pull_request:
 
 jobs:
-  run_tests:
+  build_docs:
     if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.repository
     runs-on: ubuntu-latest
     strategy:
