+++ .github/workflows/testing.yml
@@ -7,7 +7,7 @@ on:
     - cron: '00 12 * * *'  # daily at 4AM
 
 jobs:
-  build_docs:
+  run_tests:
     if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.repository
     runs-on: ubuntu-latest
     strategy:
